feat(ui): new create connection page
