+++ packages/types/lib/integration/api.ts
@@ -47,6 +47,16 @@ export type DeletePublicIntegration = Endpoint<{
     Success: { success: true };
 }>;
 
+export type ApiIntegration = Omit<Merge<IntegrationConfig, ApiTimestamps>, 'oauth_client_secret_iv' | 'oauth_client_secret_tag'>;
+
+export type GetIntegrations = Endpoint<{
+    Method: 'GET';
+    Path: '/api/v1/integrations';
+    Success: {
+        data: ApiIntegration[];
+    };
+}>;
+
 export type PostIntegration = Endpoint<{
     Method: 'POST';
     Path: '/api/v1/integrations';
@@ -57,7 +67,6 @@ export type PostIntegration = Endpoint<{
     };
 }>;
 
-export type ApiIntegration = Omit<Merge<IntegrationConfig, ApiTimestamps>, 'oauth_client_secret_iv' | 'oauth_client_secret_tag'>;
 export type GetIntegration = Endpoint<{
     Method: 'GET';
     Path: '/api/v1/integrations/:providerConfigKey';
