+++ packages/webapp/package.json
@@ -39,6 +39,7 @@
         "@nangohq/server": "file:../server",
         "@nangohq/types": "file:../types",
         "@radix-ui/react-avatar": "1.1.1",
+        "@radix-ui/react-collapsible": "1.1.1",
         "@radix-ui/react-dialog": "1.1.1",
         "@radix-ui/react-dropdown-menu": "2.1.1",
         "@radix-ui/react-hover-card": "1.1.1",
