+++ packages/webapp/src/components/SimpleTooltip.tsx
@@ -1,7 +1,12 @@
 import type React from 'react';
 import { Tooltip, TooltipProvider, TooltipContent, TooltipTrigger } from './ui/Tooltip';
+import type { Content } from '@radix-ui/react-tooltip';
 
-export const SimpleTooltip: React.FC<React.PropsWithChildren<{ tooltipContent: React.ReactNode }>> = ({ tooltipContent, children }) => {
+export const SimpleTooltip: React.FC<React.PropsWithChildren<{ tooltipContent: React.ReactNode } & React.ComponentPropsWithoutRef<typeof Content>>> = ({
+    tooltipContent,
+    children,
+    ...rest
+}) => {
     if (!tooltipContent) {
         return <>{children}</>;
     }
@@ -9,7 +14,7 @@ export const SimpleTooltip: React.FC<React.PropsWithChildren<{ tooltipContent: R
     return (
         <TooltipProvider>
             <Tooltip>
-                <TooltipContent>{tooltipContent}</TooltipContent>
+                <TooltipContent {...rest}>{tooltipContent}</TooltipContent>
                 <TooltipTrigger>{children}</TooltipTrigger>
             </Tooltip>
         </TooltipProvider>
