+++ packages/webapp/src/components/ui/Command.tsx
@@ -34,7 +34,7 @@ const CommandDialog = ({ children, ...props }: CommandDialogProps) => {
 const CommandInput = React.forwardRef<React.ElementRef<typeof CommandPrimitive.Input>, React.ComponentPropsWithoutRef<typeof CommandPrimitive.Input>>(
     ({ className, ...props }, ref) => (
         // eslint-disable-next-line react/no-unknown-property
-        <div className="flex items-center border-b px-3" cmdk-input-wrapper="">
+        <div className="flex items-center" cmdk-input-wrapper="">
             <CommandPrimitive.Input
                 ref={ref}
                 className={cn(
